we have integer channel ids
git-svn-id: http://red5.googlecode.com/svn/java/server/trunk@2193 1b6495e4-3631-0410-8e05-8f51eee8b9cc

+++ src/org/red5/server/api/service/IServiceCapableConnection.java
@@ -37,7 +37,7 @@ public interface IServiceCapableConnection extends IConnection {
      * @param call       Service call
      * @param channel    Channel used
      */
-    void invoke(IServiceCall call, byte channel);
+    void invoke(IServiceCall call, int channel);
 
     /**
      * Invoke method by name
@@ -79,7 +79,7 @@ public interface IServiceCapableConnection extends IConnection {
      * @param call
      * @param channel
      */
-    void notify(IServiceCall call, byte channel);
+    void notify(IServiceCall call, int channel);
 
     /**
      *
